pyproject.toml
[tool.poetry]
name = "cygnet-energy"
version = "0.1.0"
description = "European grid intelligence platform - Real-time electricity system data from ENTSO-E"
authors = ["Yashar Naghdi <he@yasharnaghdi.com>"]
readme = "README.md"
repository = "https://github.com/yasharnaghdi/cygnet-energy"
keywords = ["grid", "electricity", "entso-e", "energy", "timeseries", "carbon"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {version = "^0.27.0", extras = ["standard"]}
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
apscheduler = "^3.10.4"
prometheus-client = "^0.19.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
pytz = "^2024.1"
pandas = "^2.1.0"
numpy = "^1.26.0"
streamlit = "^1.29.0"
plotly = "^5.18.0"
lxml = "^5.0.0"
scikit-learn = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
